# Simplified push method; updated first and max slice sizes