… testdata/stack.jpg